backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Mail==0.9.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.0.1

# Security
Werkzeug==3.0.6
cryptography
bcrypt==4.1.2
PyJWT==2.8.0

# Validation and serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
pydantic==2.5.2

# Async tasks
celery==5.3.4
kombu==5.3.4

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==24.3.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Documentation
flask-restx==1.3.0
flasgger==0.9.7.1

# Environment management
python-dotenv==1.0.0

# Financial and compliance
python-dateutil==2.8.2


# Utilities
requests==2.32.4
click==8.1.7
gunicorn==22.0.0
